WIP: refacto
