+++ smashtheque/helpers/misc.py
@@ -1,4 +1,6 @@
+from collections import UserDict
 from collections.abc import Mapping
+import re
 import unicodedata
 
 def is_discord_id(v):
@@ -24,3 +26,18 @@ class Map(UserDict):
     if isinstance(val, Mapping):
         return Map(val)
     return val
+
+def format_emoji(emoji_id):
+  return f"<:placeholder:{emoji_id}>"
+
+def format_character(character):
+  return format_emoji(character["emoji"])
+
+def format_discord_user(discord_id):
+  return f"<@{discord_id}>"
+
+def format_team(team):
+  return "{0} ({1})".format(team["name"], team["short_name"])
+
+def format_location(location):
+  return location["name"].title()
