+++ smashtheque/smashtheque.py
@@ -1,251 +1,21 @@
-from redbot.core import commands
 from redbot.core.utils.predicates import ReactionPredicate
-from redbot.core.utils.menus import start_adding_reactions
-from redbot.core.utils.chat_formatting import humanize_list
 from redbot.core.utils.predicates import MessagePredicate
 
 import discord
 import asyncio
-import aiohttp
-from redbot.core.bot import Red
-from typing import Optional
-import re
 import json
 import math
 import rollbar
 import os
 import sys
 
-from collections import UserDict
 from random import choice
 
-import misc
-
-async def yeet(ctx, erreur):
-    """lever des erreurs"""
-    embed = discord.Embed(
-        title=f"Erreur dans la commande {ctx.command} :",
-        description=erreur,
-        colour=discord.Colour.red()
-    )
-    embed.set_author(
-        name="Smashthèque",
-        icon_url="https://cdn.discordapp.com/avatars/745022618356416572/c8fa739c82cdc5a730d9bdf411a552b0.png?size=1024",
-    )
-    await ctx.send(embed=embed)
-
-async def uniqueyeet(ctx, erreur, playername):
-    """lever des erreurs en ajoutant le nom du joueur"""
-    global failed_lines
-    failed_lines.append(playername)
-    embed = discord.Embed(
-        title=f"Erreur pendant le traitement de la ligne {playername} :",
-        description=erreur,
-        colour=discord.Colour.red()
-    )
-    embed.set_author(
-        name="Smashthèque",
-        icon_url="https://cdn.discordapp.com/avatars/745022618356416572/c8fa739c82cdc5a730d9bdf411a552b0.png?size=1024",
-    )
-    await ctx.send(embed=embed)
-
-async def generic_error(ctx, error):
-    await yeet(ctx, "T'as cassé le bot, GG. Tu peux contacter <@332894758076678144> ou <@608210202952466464> s'il te plaît ?")
-    rollbar.report_exc_info(sys.exc_info(), error)
-
-def format_emoji(emoji_id):
-    return f"<:placeholder:{emoji_id}>"
-
-def format_character(character):
-    return format_emoji(character["emoji"])
-
-def format_discord_user(discord_id):
-    return f"<@{discord_id}>"
-
-def format_team(team):
-    return "{0} ({1})".format(team["name"], team["short_name"])
-
-def format_location(location):
-    return location["name"].title()
-
-class Error(Exception):
-    """Base class for other exceptions"""
-    pass
-
-class CharactersCacheNotFetched(Error):
-    """Raised when the cache is not fetched yet"""
-    pass
+from .helpers.misc import *
+from .helpers.dialogs import *
 
 
-class Smashtheque(commands.Cog):
-    async def initialize(self):
-        if 'ROLLBAR_TOKEN' in os.environ and os.environ['ROLLBAR_TOKEN']:
-            rollbar_token = os.environ['ROLLBAR_TOKEN']
-            if 'ROLLBAR_ENV' in os.environ and os.environ['ROLLBAR_ENV']:
-                rollbar_env = os.environ['ROLLBAR_ENV']
-            else:
-                rollbar_env = 'production'
-
-        else:
-            rollbar_token = await self.bot.get_shared_api_tokens("smashtheque")
-            rollbar_token = rollbar_token["token"]
-            if 'environment' in rollbar_token:
-                rollbar_env = rollbar_token["environment"]
-            else:
-                rollbar_env = 'production'
-        rollbar.init(rollbar_token, rollbar_env)
-
-        try:
-            if 'SMASHTHEQUE_API_URL' in os.environ and os.environ['SMASHTHEQUE_API_URL']:
-                self.api_base_url = os.environ['SMASHTHEQUE_API_URL']
-            else:
-                api_base_url = await self.bot.get_shared_api_tokens("smashtheque")
-                self.api_base_url = api_base_url["url"]
-            print(f"Smashthèque API base URL set to {self.api_base_url}")
-            if 'SMASHTHEQUE_API_TOKEN' in os.environ and os.environ['SMASHTHEQUE_API_TOKEN']:
-                bearer = os.environ['SMASHTHEQUE_API_TOKEN']
-            else:
-                bearer = await self.bot.get_shared_api_tokens("smashtheque")
-                bearer = bearer["bearer"]
-
-            self.api = ApiClient(apiBaseUrl=api_base_url, bearerToken=bearer)
-        except:
-            rollbar.report_exc_info()
-            raise
-
-    def __init__(self, bot: Red):
-        self.bot = bot
-
-    def cog_unload(self):
-        asyncio.create_task(self._session.close())
-
-
-
-
-
-
-
-    async def find_team_by_short_name(self, short_name):
-        request_url = "{0}?by_short_name_like={1}".format(self.api_url("teams"), short_name)
-        async with self._session.get(request_url) as response:
-            teams = await response.json()
-            if teams != []:
-                # puts values in cache before responding
-                for team in teams:
-                    self._teams_cache[str(team["id"])] = team
-                return teams[0]
-            else:
-                return None
-
-    async def find_team_by_id(self, team_id):
-        request_url = "{api_url}/{team_id}".format(api_url=self.api_url("teams"), team_id=team_id)
-        async with self._session.get(request_url) as response:
-            team = await response.json()
-            return team #or team[0]
-
-    async def find_tournament_by_id(self, tournament_id):
-        request_url = f"{self.api_url('recurring_tournaments')}/{tournament_id}"
-        async with self._session.get(request_url) as response:
-            tournament = await response.json()
-            return tournament
-
-    async def find_location_by_name(self, name):
-        request_url = "{0}?by_name_like={1}".format(self.api_url("locations"), name)
-        async with self._session.get(request_url) as response:
-            locations = await response.json()
-            if locations != []:
-                # puts values in cache before responding
-                for location in locations:
-                    self._locations_cache[str(location["id"])] = location
-                return locations[0]
-            else:
-                return None
-
-    async def find_player_by_id(self, player_id):
-        request_url = "{0}/{1}".format(self.api_url("players"), player_id)
-        async with self._session.get(request_url) as response:
-            player = await response.json()
-            return player
-
-    async def find_player_by_ids(self, player_ids):
-        players = []
-        for player_id in player_ids:
-            player = await self.find_player_by_id(player_id)
-            players.append(player)
-        return players
-
-    async def find_player_by_discord_id(self, discord_id):
-        request_url = "{0}?by_discord_id={1}".format(self.api_url("players"), discord_id)
-        async with self._session.get(request_url) as response:
-            players = await response.json()
-            if len(players) > 0:
-                return players[0]
-            return None
-
-    async def find_players_by_name_like(self, name):
-        request_url = "{0}?by_name_like={1}".format(self.api_url("players"), name)
-        async with self._session.get(request_url) as response:
-            players = await response.json()
-            return players
-
-    async def find_member_by_discord_id(self, discord_id):
-        request_url = "{api_url}/{discord_id}".format(api_url=self.api_url("discord_users"), discord_id=discord_id)
-        async with self._session.get(request_url) as response:
-            player = await response.json()
-            return player if player != [] else None
-
-    async def raise_message(self, ctx, message):
-        embed = discord.Embed(title=message)
-        embed.set_author(
-            name="Smashthèque",
-            icon_url="https://cdn.discordapp.com/avatars/745022618356416572/c8fa739c82cdc5a730d9bdf411a552b0.png?size=1024",
-        )
-        await ctx.send(embed=embed)
-
-    async def raise_not_linked(self, ctx):
-        await self.raise_message(ctx, f"Votre compte Discord n'est associé à aucun joueur.\nUtilisez `{ctx.clean_prefix}jesuis` pour associer votre compte à un joueur.")
-
-    async def ask_confirmation(self, ctx, embed):
-        temp_message = await ctx.send(embed=embed)
-        pred = ReactionPredicate.yes_or_no(temp_message, ctx.author)
-        start_adding_reactions(
-            temp_message, ReactionPredicate.YES_OR_NO_EMOJIS
-        )
-        try:
-            await self.bot.wait_for("reaction_add", timeout=60.0, check=pred)
-        except asyncio.TimeoutError:
-            await ctx.send("Commande annulée")
-            return False
-        if pred.result is True:
-            await temp_message.delete()
-            return True
-        else:
-            await ctx.send("Commande annulée")
-            await temp_message.delete()
-            return False
-
-    async def ask_choice(self, ctx, embed, elements_count):
-        temp_message = await ctx.send(embed=embed)
-        emojis = ReactionPredicate.NUMBER_EMOJIS[1:elements_count + 1]
-        start_adding_reactions(temp_message, emojis)
-        pred = ReactionPredicate.with_emojis(emojis, temp_message)
-        try:
-            await ctx.bot.wait_for("reaction_add", timeout=60.0, check=pred)
-        except asyncio.TimeoutError:
-            await ctx.send("Commande annulée")
-            return None
-        if type(pred.result) == int:
-            await temp_message.delete()
-            return pred.result
-        return None
-
-    async def show_confirmation(self, ctx, message):
-        embed = discord.Embed(
-            title="I guess it's done!",
-            description=message,
-            colour=discord.Colour.green()
-        )
-        await ctx.send(embed=embed)
+class Smashtheque:
 
     def embed_players(self, embed, players, with_index=False):
         idx = 0
@@ -264,7 +34,7 @@ class Smashtheque(commands.Cog):
                 personnages.append(format_character(character))
         elif "character_ids" in _player:
             for character_id in player.character_ids:
-                personnages.append(format_character(self._characters_cache[str(character_id)]))
+                personnages.append(format_character(self.api._characters_cache[str(character_id)]))
         if len(personnages) < 1:
             personnages.append("\u200b")
 
@@ -279,7 +49,7 @@ class Smashtheque(commands.Cog):
                 team_names.append(format_team(team))
         elif "team_ids" in _player:
             for team_id in player.team_ids:
-                team_names.append(format_team(self._teams_cache[str(team_id)]))
+                team_names.append(format_team(self.api._teams_cache[str(team_id)]))
         if len(team_names) > 0:
             embed.add_field(name="Équipes", value="\n".join(team_names), inline=True)
 
@@ -289,7 +59,7 @@ class Smashtheque(commands.Cog):
                 location_names.append(format_location(location))
         elif "location_ids" in _player:
             for location_id in player.location_ids:
-                location_names.append(format_location(self._locations_cache[str(location_id)]))
+                location_names.append(format_location(self.api._locations_cache[str(location_id)]))
         if len(location_names) > 0:
             embed.add_field(name="Localisations", value="\n".join(location_names), inline=True)
 
@@ -304,76 +74,75 @@ class Smashtheque(commands.Cog):
         )
         embed.set_footer(text="Réagissez avec ✅ pour confirmer et créer ce joueur, ou réagissez avec ❎ pour annuler.")
         self.embed_player(embed, player)
-        doit = await self.ask_confirmation(ctx, embed)
+        doit = await ask_confirmation(ctx, embed)
         if doit:
             await self.create_player(ctx, player)
 
     async def create_player(self, ctx, player):
         print(f"create player {player}")
-        payload = {"player": player}
-        async with self._session.post(self.api_url("players"), json=payload) as r:
-            if r.status == 201:
-                # player creation wen fine
-                player_name = player["name"]
-                await self.show_confirmation(ctx, f"Le joueur {player_name} a été ajouté à la Smashthèque et est en attente de validation.")
-                return
+        r = await self.api.createPlayer(player)
+        if r.status == 201:
+            # player creation wen fine
+            player_name = player["name"]
+            await show_confirmation(ctx, f"Le joueur {player_name} a été ajouté à la Smashthèque et est en attente de validation.")
+            return
 
-            if r.status == 422:
-                result = await r.json()
-                erreur = Map(result)
-                print(f"errors: {erreur.errors}")
-                if "name" in erreur.errors and erreur.errors["name"] == "already_known":
-                    alts = await self.find_player_by_ids(erreur.errors["existing_ids"])
-                    embed = discord.Embed(
-                        title="Un ou plusieurs joueurs possèdent le même pseudo que le joueur que vous souhaitez ajouter.",
-                        colour=discord.Colour.blue()
-                    )
-                    embed.set_footer(text="Réagissez avec ✅ pour confirmer et créer un nouveau joueur, ou\nréagissez avec ❎ pour annuler.")
-                    self.embed_players(embed, alts)
-                    doit = await self.ask_confirmation(ctx, embed)
-                    if doit:
-                        player["name_confirmation"] = True
-                        await self.create_player(ctx, player)
-                elif "discord_user" in erreur.errors and erreur.errors["discord_user"] == ["already_taken"]:
-                    await yeet(ctx, "Ce compte Discord est déjà relié à un autre joueur dans la Smashthèque.")
-                    return
-                else:
-                    await generic_error(ctx, erreur)
-                    return
+        if r.status == 422:
+            result = await r.json()
+            erreur = Map(result)
+            print(f"errors: {erreur.errors}")
+            if "name" in erreur.errors and erreur.errors["name"] == "already_known":
+                alts = await self.api.findPlayerByIds(erreur.errors["existing_ids"])
+                embed = discord.Embed(
+                    title="Un ou plusieurs joueurs possèdent le même pseudo que le joueur que vous souhaitez ajouter.",
+                    colour=discord.Colour.blue()
+                )
+                embed.set_footer(text="Réagissez avec ✅ pour confirmer et créer un nouveau joueur, ou\nréagissez avec ❎ pour annuler.")
+                self.embed_players(embed, alts)
+                doit = await ask_confirmation(ctx, embed)
+                if doit:
+                    player["name_confirmation"] = True
+                    await self.create_player(ctx, player)
+            elif "discord_user" in erreur.errors and erreur.errors["discord_user"] == ["already_taken"]:
+                await yeet(ctx, "Ce compte Discord est déjà relié à un autre joueur dans la Smashthèque.")
+                return
+            else:
+                await generic_error(ctx)
+                rollbar.report_exc_info(sys.exc_info(), erreur)
+                return
 
     async def update_player(self, ctx, player_id, data):
         print(f"update player {player_id} with {data}")
-        payload = {"player": data}
-        player_url = "{0}/{1}".format(self.api_url("players"), player_id)
-        async with self._session.patch(player_url, json=payload) as r:
-            if r.status == 200:
-                embed = discord.Embed(title="Joueur mis à jour :")
-                self.embed_player(embed, await r.json())
-                await ctx.send(embed=embed)
-                return
+        r = await self.api.updatePlayer(player_id, data)
+        if r.status == 200:
+            embed = discord.Embed(title="Joueur mis à jour :")
+            self.embed_player(embed, await r.json())
+            await ctx.send(embed=embed)
+            return
 
-            if r.status == 422:
-                result = await r.json()
-                erreur = Map(result)
-                print(f"errors: {erreur.errors}")
-                if "name" in erreur.errors and erreur.errors["name"] == "already_known":
-                    alts = await self.find_player_by_ids(erreur.errors["existing_ids"])
-                    embed = discord.Embed(
-                        title="Un ou plusieurs autres joueurs utilisent ce pseudo.",
-                        colour=discord.Colour.blue()
-                    )
-                    embed.set_footer(text="Réagissez avec ✅ pour confirmer et mettre à jour, ou\nréagissez avec ❎ pour annuler.")
-                    self.embed_players(embed, alts)
-                    doit = await self.ask_confirmation(ctx, embed)
-                    if doit:
-                        data["name_confirmation"] = True
-                        await self.update_player(ctx, player_id, data)
-                elif "discord_user" in erreur.errors and erreur.errors["discord_user"] == ["already_taken"]:
-                    await yeet(ctx, "Ce compte Discord est déjà relié à un autre joueur dans la Smashthèque.")
-                    return
-                else:
-                    await generic_error(ctx, erreur)
-                    return
+        if r.status == 422:
+            result = await r.json()
+            erreur = Map(result)
+            print(f"errors: {erreur.errors}")
+            if "name" in erreur.errors and erreur.errors["name"] == "already_known":
+                alts = await self.api.findPlayerByIds(erreur.errors["existing_ids"])
+                embed = discord.Embed(
+                    title="Un ou plusieurs autres joueurs utilisent ce pseudo.",
+                    colour=discord.Colour.blue()
+                )
+                embed.set_footer(text="Réagissez avec ✅ pour confirmer et mettre à jour, ou\nréagissez avec ❎ pour annuler.")
+                self.embed_players(embed, alts)
+                doit = await ask_confirmation(ctx, embed)
+                if doit:
+                    data["name_confirmation"] = True
+                    await self.update_player(ctx, player_id, data)
+            elif "discord_user" in erreur.errors and erreur.errors["discord_user"] == ["already_taken"]:
+                await yeet(ctx, "Ce compte Discord est déjà relié à un autre joueur dans la Smashthèque.")
+                return
+            else:
+                await generic_error(ctx)
+                rollbar.report_exc_info(sys.exc_info(), erreur)
+                return
 
     async def complete_bracket_link(self, ctx, tournament):
 
@@ -394,28 +163,32 @@ class Smashtheque(commands.Cog):
         await ctx.send("Si votre tournois possède un graph, veuillez réutiliser la même commande avec le lien du tournois comme argument, et le graph comme attachement.")
         return
 
+
+
+
+
+
+
     async def do_createlocation(self, ctx, name, country=False):
         print(f"create location {name}")
-        payload = {"name": name}
-        if country:
-            payload["type"] = "Locations::Country"
-        async with self._session.post(self.api_url("locations"), json=payload) as r:
-            if r.status == 201:
-                # location creation went fine
-                await self.show_confirmation(ctx, f"La localisation {name} a été ajoutée à la base de données.")
+        r = await self.api.createLocation(name, country=country)
+        if r.status == 201:
+            # location creation went fine
+            await show_confirmation(ctx, f"La localisation {name} a été ajoutée à la base de données.")
+            return
+
+        if r.status == 422:
+            result = await r.json()
+            erreur = Map(result)
+            print(erreur.errors["name"])
+            if erreur.errors["name"] == ["not_unique"]:
+                await yeet(ctx, "Cette localisation existe déjà dans la Smashthèque.")
                 return
 
-            if r.status == 422:
-                result = await r.json()
-                erreur = Map(result)
-                print(erreur.errors["name"])
-                if erreur.errors["name"] == ["not_unique"]:
-                    await yeet(ctx, "Cette localisation existe déjà dans la Smashthèque.")
-                    return
-
-            # something went wrong but we don't know what
-            await generic_error(ctx, r)
-            return
+        # something went wrong but we don't know what
+        await generic_error(ctx)
+        rollbar.report_exc_info(sys.exc_info(), r)
+        return
 
     async def do_addplayer(self, ctx, arg):
 
@@ -476,7 +249,7 @@ class Smashtheque(commands.Cog):
                 # at this stage, the next argument could be a character emoji
                 # ... [emoji] [emoji] [team] [location] [discord ID]
 
-                if not self.isCharacter(argu):
+                if not self.api.isCharacter(argu):
                     # we are actually done with the emojis, so go to stage 2
                     current_stage = 2
                     # do not 'continue' here because we stil want to process @argu
@@ -502,7 +275,7 @@ class Smashtheque(commands.Cog):
                     response["discord_id"] = str(argu)
                     break
 
-                team = await self.find_team_by_short_name(argu)
+                team = await self.api.findTeamByShortName(argu)
                 if team != None:
                     response["team_ids"].append(team["id"])
                     current_stage = 3
@@ -522,7 +295,7 @@ class Smashtheque(commands.Cog):
                     response["discord_id"] = str(argu)
                     break
 
-                location = await self.find_location_by_name(argu)
+                location = await self.api.findLocationByName(argu)
                 if location != None:
                     response["location_ids"].append(location["id"])
                     current_stage = 4
@@ -583,18 +356,18 @@ class Smashtheque(commands.Cog):
         await self.confirm_create_player(ctx, response)
 
     async def do_link(self, ctx, pseudo, discord_id):
-        player = await self.find_player_by_discord_id(discord_id)
+        player = await self.api.findPlayerByDiscordId(discord_id)
         if player != None:
             embed = discord.Embed(title="Un joueur est déjà associé à ce compte Discord. Contactez un admin pour dissocier ce compte Discord de ce joueur.")
             self.embed_player(embed, player)
             await ctx.send(embed=embed)
             return
 
-        players = await self.find_players_by_name_like(pseudo)
+        players = await self.api.findPlayersByNameLike(pseudo)
 
         # no player found
         if len(players) == 0:
-            await self.raise_message(ctx, f"Ce pseudo n'existe pas.\nUtilisez la commande `{ctx.clean_prefix}creerjoueur` pour l'ajouter.\n")
+            await raise_message(ctx, f"Ce pseudo n'existe pas.\nUtilisez la commande `{ctx.clean_prefix}creerjoueur` pour l'ajouter.\n")
             return
 
         # one player found: ask confirmation
@@ -605,15 +378,15 @@ class Smashtheque(commands.Cog):
                 colour=discord.Colour.blue()
             )
             self.embed_player(embed, player)
-            doit = await self.ask_confirmation(ctx, embed)
+            doit = await ask_confirmation(ctx, embed)
             if not doit:
                 return
             if player["discord_id"] != None:
-                await self.raise_message(ctx, "Il semblerait que ce joueur soit déjà associé à un compte Discord")
+                await raise_message(ctx, "Il semblerait que ce joueur soit déjà associé à un compte Discord")
                 return
             await self.update_player(ctx, player["id"], {"discord_id": str(discord_id)})
             discord_user = format_discord_user(discord_id)
-            await self.show_confirmation(ctx, f"Le compte Discord {discord_user} est maintenant associé au joueur {pseudo}.")
+            await show_confirmation(ctx, f"Le compte Discord {discord_user} est maintenant associé au joueur {pseudo}.")
             return
 
         # multiple players found: ask which one
@@ -622,37 +395,37 @@ class Smashtheque(commands.Cog):
             colour=discord.Colour.blue()
         )
         self.embed_players(embed, players, with_index=True)
-        choice = await self.ask_choice(ctx, embed, len(players))
+        choice = await ask_choice(ctx, embed, len(players))
         player = players[choice]
         if player["discord_id"] != None:
-            await self.raise_message(ctx, "Il semblerait que ce joueur soit déjà associé à un compte Discord")
+            await raise_message(ctx, "Il semblerait que ce joueur soit déjà associé à un compte Discord")
             return
         await self.update_player(ctx, player["id"], {"discord_id": str(discord_id)})
         discord_user = format_discord_user(discord_id)
-        await self.show_confirmation(ctx, f"Le compte Discord {discord_user} est maintenant associé au joueur {pseudo}.")
+        await show_confirmation(ctx, f"Le compte Discord {discord_user} est maintenant associé au joueur {pseudo}.")
 
     async def do_unlink(self, ctx, target_member):
         discord_id = target_member.id
-        player = await self.find_player_by_discord_id(discord_id)
+        player = await self.api.findPlayerByDiscordId(discord_id)
         if player == None:
-            await self.raise_not_linked(ctx)
+            await raise_not_linked(ctx)
             return
         embed = discord.Embed(title="Un joueur est associé à ce compte discord. Voulez vous le dissocier ?")
         embed.set_footer(text="Réagissez avec ✅ pour confirmer et dissocier ce compte du joueur, ou\nréagissez avec ❎ pour annuler.")
         self.embed_player(embed, player)
-        doit = await self.ask_confirmation(ctx, embed)
+        doit = await ask_confirmation(ctx, embed)
         if not doit:
             return
         await self.update_player(ctx, player["id"], {"discord_id": None})
         player_name = player["name"]
         discord_user = format_discord_user(discord_id)
-        await self.show_confirmation(ctx, f"Le compte Discord {discord_user} a été dissocié du joueur {player_name}.")
+        await show_confirmation(ctx, f"Le compte Discord {discord_user} a été dissocié du joueur {player_name}.")
 
     async def do_showplayer(self, ctx, target_member):
         discord_id = target_member.id
-        player = await self.find_player_by_discord_id(discord_id)
+        player = await self.api.findPlayerByDiscordId(discord_id)
         if player == None:
-            await self.raise_not_linked(ctx)
+            await raise_not_linked(ctx)
             return
         embed = discord.Embed(
             title="Ce compte Discord est associé au joueur suivant :",
@@ -662,11 +435,11 @@ class Smashtheque(commands.Cog):
         await ctx.send(embed=embed)
 
     async def do_findplayer(self, ctx, name):
-        players = await self.find_players_by_name_like(name)
+        players = await self.api.findPlayersByNameLike(name)
 
         # no player found
         if len(players) == 0:
-            await self.raise_message(ctx, f"Aucun joueur connu avec ce pseudo.\nVous pouvez utiliser la commande `{ctx.clean_prefix}creerjoueur` pour ajouter un joueur.\n")
+            await raise_message(ctx, f"Aucun joueur connu avec ce pseudo.\nVous pouvez utiliser la commande `{ctx.clean_prefix}creerjoueur` pour ajouter un joueur.\n")
             return
 
         if len(players) == 1:
@@ -686,21 +459,21 @@ class Smashtheque(commands.Cog):
         return
 
     async def do_editname(self, ctx, discord_id, new_name):
-        player = await self.find_player_by_discord_id(discord_id)
+        player = await self.api.findPlayerByDiscordId(discord_id)
         if player == None:
-            await self.raise_not_linked(ctx)
+            await raise_not_linked(ctx)
             return
         await self.update_player(ctx, player["id"], {"name": new_name})
 
     async def do_removelocation(self, ctx, discord_id, location_name):
-        player = await self.find_player_by_discord_id(discord_id)
+        player = await self.api.findPlayerByDiscordId(discord_id)
         if player == None:
-            await self.raise_not_linked(ctx)
+            await raise_not_linked(ctx)
             return
         location_ids = player["location_ids"]
-        location = await self.find_location_by_name(location_name)
+        location = await self.api.findLocationByName(location_name)
         if location == None:
-            await self.raise_message(
+            await raise_message(
                 ctx,
                 f"Nous n'avons pas réussi à trouver {location_name}.\nS'il s'agit d'une ville, vous pouvez l'ajouter à la Smashthèque avec `{ctx.clean_prefix}creerville`.\nS'il s'agit d'un pays, vous pouvez l'ajouter à la Smashthèque avec `{ctx.clean_prefix}creerpays`"
             )
@@ -711,21 +484,21 @@ class Smashtheque(commands.Cog):
         await self.update_player(ctx, player["id"], {"location_ids": location_ids})
 
     async def do_addlocation(self, ctx, discord_id, location_name):
-        player = await self.find_player_by_discord_id(discord_id)
+        player = await self.api.findPlayerByDiscordId(discord_id)
         if player == None:
-            await self.raise_not_linked(ctx)
+            await raise_not_linked(ctx)
             return
         location_ids = player["location_ids"]
-        location = await self.find_location_by_name(location_name)
+        location = await self.api.findLocationByName(location_name)
         if location == None:
-            await self.raise_message(
+            await raise_message(
                 ctx,
                 f"Nous n'avons pas réussi à trouver {location_name}.\nS'il s'agit d'une ville, vous pouvez l'ajouter à la Smashthèque avec `{ctx.clean_prefix}creerville`.\nS'il s'agit d'un pays, vous pouvez l'ajouter à la Smashthèque avec `{ctx.clean_prefix}creerpays`."
             )
             return
         location_id = location["id"]
         if location_id in location_ids:
-            await self.raise_message(
+            await raise_message(
                 ctx,
                 f"Ce joueur est déjà localisé à {location_name}."
             )
@@ -734,14 +507,14 @@ class Smashtheque(commands.Cog):
         await self.update_player(ctx, player["id"], {"location_ids": location_ids})
 
     async def do_removeteam(self, ctx, discord_id, team_short_name):
-        player = await self.find_player_by_discord_id(discord_id)
+        player = await self.api.findPlayerByDiscordId(discord_id)
         if player == None:
-            await self.raise_not_linked(ctx)
+            await raise_not_linked(ctx)
             return
         team_ids = player["team_ids"]
-        team = await self.find_team_by_short_name(team_short_name)
+        team = await self.api.findTeamByShortName(team_short_name)
         if team == None:
-            await self.raise_message(
+            await raise_message(
                 ctx,
                 f"Nous n'avons pas réussi à trouver l'équipe {team_short_name}.\nVous pouvez demander à un administrateur de la créer."
             )
@@ -752,21 +525,21 @@ class Smashtheque(commands.Cog):
         await self.update_player(ctx, player["id"], {"team_ids": team_ids})
 
     async def do_addteam(self, ctx, discord_id, team_short_name):
-        player = await self.find_player_by_discord_id(discord_id)
+        player = await self.api.findPlayerByDiscordId(discord_id)
         if player == None:
-            await self.raise_not_linked(ctx)
+            await raise_not_linked(ctx)
             return
         team_ids = player["team_ids"]
-        team = await self.find_team_by_short_name(team_short_name)
+        team = await self.api.findTeamByShortName(team_short_name)
         if team == None:
-            await self.raise_message(
+            await raise_message(
                 ctx,
                 f"Nous n'avons pas réussi à trouver l'équipe {team_short_name}.\nVous pouvez demander à un administrateur de la créer."
             )
             return
         team_id = team["id"]
         if team_id in team_ids:
-            await self.raise_message(
+            await raise_message(
                 ctx,
                 f"Ce joueur est déjà membre de l'équipe {team_short_name}."
             )
@@ -780,8 +553,8 @@ class Smashtheque(commands.Cog):
         await self.api.initCache()
 
         lines = []
-        for character_id in self._characters_cache:
-            character = self._characters_cache[str(character_id)]
+        for character_id in self.api._characters_cache:
+            character = self.api._characters_cache[str(character_id)]
             tag = format_character(character)
             name = character["name"]
             lines.append([tag, name])
@@ -809,9 +582,9 @@ class Smashtheque(commands.Cog):
         # fill characters cache if empty
         await self.api.initCache()
 
-        player = await self.find_player_by_discord_id(discord_id)
+        player = await self.api.findPlayerByDiscordId(discord_id)
         if player == None:
-            await self.raise_not_linked(ctx)
+            await raise_not_linked(ctx)
             return
         character_ids = player["character_ids"]
         for label in labels.split():
@@ -822,7 +595,7 @@ class Smashtheque(commands.Cog):
             character_id = character["id"]
             if character_id in character_ids:
                 character_tag = format_character(character)
-                await self.raise_message(
+                await raise_message(
                     ctx,
                     f"{character_tag} est déjà indiqué sur ce joueur."
                 )
@@ -835,9 +608,9 @@ class Smashtheque(commands.Cog):
         # fill characters cache if empty
         await self.api.initCache()
 
-        player = await self.find_player_by_discord_id(discord_id)
+        player = await self.api.findPlayerByDiscordId(discord_id)
         if player == None:
-            await self.raise_not_linked(ctx)
+            await raise_not_linked(ctx)
             return
         character_ids = player["character_ids"]
         for label in labels.split():
@@ -849,7 +622,7 @@ class Smashtheque(commands.Cog):
             if char_id in character_ids:
                 character_ids.remove(char_id)
         if len(character_ids) < 1:
-            await self.raise_message(ctx, "Un joueur doit jouer au moins un perso")
+            await raise_message(ctx, "Un joueur doit jouer au moins un perso")
             return
         await self.update_player(ctx, player["id"], {"character_ids": character_ids})
 
@@ -858,9 +631,9 @@ class Smashtheque(commands.Cog):
         # fill characters cache if empty
         await self.api.initCache()
 
-        player = await self.find_player_by_discord_id(discord_id)
+        player = await self.api.findPlayerByDiscordId(discord_id)
         if player == None:
-            await self.raise_not_linked(ctx)
+            await raise_not_linked(ctx)
             return
         character_ids = []
         for label in labels.split():
@@ -876,11 +649,12 @@ class Smashtheque(commands.Cog):
         await ctx.send(result)
         return
 
+
     async def maj_team_infos(self, ctx, object_name):
 
         """will update a team info
         object_name is something like logo, or roster"""
-        player = await self.find_member_by_discord_id(ctx.author.id)
+        player = await self.api.findDiscordUserByDiscordId(ctx.author.id)
         if player is None:
             await yeet(ctx, "Vous n'êtes pas enregistré dans la smashtheque.")
             return
@@ -895,40 +669,38 @@ class Smashtheque(commands.Cog):
             embed = discord.Embed(title=f"Mise à jour du {object_name} de votre team", description=f"Pour changer le {object_name} de votre team, veuillez utiliser cette commande avec une image comme attachement", colour= await ctx.embed_colour())
             await ctx.send(embed=embed)
             return
-        team = await self.find_team_by_id(player["administrated_teams"][0]["id"])
+        team = await self.api.findTeamById(player["administrated_teams"][0]["id"])
         if len(player["administrated_teams"]) > 1:
             embed = discord.Embed(title="Vous âtes administrateur de plusieurs teams. ", description=f"De quelle team faut-il modifier le {object_name} ?")
             for team_entry in player["administrated_teams"]:
                 embed.add_field(name=team_entry["short_name"], value=team_entry["name"])
-            choice_result = await self.ask_choice(ctx, embed, len(player["administrated_teams"]))
-            team = await self.find_team_by_id(player["administrated_teams"][choice_result]["id"])
+            choice_result = await ask_choice(ctx, embed, len(player["administrated_teams"]))
+            team = await self.api.findTeamById(player["administrated_teams"][choice_result]["id"])
         embed = discord.Embed(title=f"Vous êtes sur le point de changer le {object_name} de la team {team['name']} pour :")
         embed.set_author(
         name="Smashthèque",
         icon_url="https://cdn.discordapp.com/avatars/745022618356416572/c8fa739c82cdc5a730d9bdf411a552b0.png?size=1024",
         )
         embed.set_image(url=ctx.message.attachments[0].url)
-        confirmation = await self.ask_confirmation(ctx, embed)
+        confirmation = await ask_confirmation(ctx, embed)
         if not confirmation:
             return
-        request_url = f"{self.api_url('teams')}/{team['id']}/"
-        reply_body = {
-            "team": {
-                f"{object_name}_url": ctx.message.attachments[0].url
-            }
+        team_data = {
+            f"{object_name}_url": ctx.message.attachments[0].url
         }
-        async with self._session.patch(request_url, json=reply_body) as response:
-            if response.status != 200:
-                await generic_error(ctx, f"error in command majlogo : status code {response.status}, response : {response}")
-                print(response)
-                print(response.text)
-                return
-            else:
-                await self.show_confirmation(ctx, f"Le {object_name} de la team {team['name']} a été mis à jour avec succès.")
+        response = await self.api.updateTeam(team['id'], team_data)
+        if response.status != 200:
+            await generic_error(ctx)
+            rollbar.report_exc_info(sys.exc_info(), f"error in command majlogo : status code {response.status}, response : {response}")
+            print(response)
+            print(response.text)
+            return
+        else:
+            await show_confirmation(ctx, f"Le {object_name} de la team {team['name']} a été mis à jour avec succès.")
 
     async def do_addedition(self, ctx, bracket):
         #generic checks
-        player = await self.find_member_by_discord_id(ctx.author.id)
+        player = await self.api.findDiscordUserByDiscordId(ctx.author.id)
         if player is None:
             await yeet(ctx, "Vous n'êtes pas enregistré dans la smashtheque.")
             return
@@ -943,13 +715,13 @@ class Smashtheque(commands.Cog):
             for tournament_entry in player["administrated_recurring_tournaments"]:
                 embed.add_field(name=(1+idx), value=tournament_entry["name"], inline=False)
                 idx += 1
-            choice = await self.ask_choice(ctx, embed, len(player["administrated_recurring_tournaments"]))
+            choice = await ask_choice(ctx, embed, len(player["administrated_recurring_tournaments"]))
             if choice == None:
                 return
-            tournament = await self.find_tournament_by_id(player["administrated_recurring_tournaments"][choice]["id"])
+            tournament = await self.api.findTournamentById(player["administrated_recurring_tournaments"][choice]["id"])
 
         else:
-            tournament = await self.find_tournament_by_id(player["administrated_recurring_tournaments"][0]["id"])
+            tournament = await self.api.findTournamentById(player["administrated_recurring_tournaments"][0]["id"])
 
         #completing the bracket link
         if not bracket:
@@ -979,251 +751,17 @@ class Smashtheque(commands.Cog):
             icon_url="https://cdn.discordapp.com/avatars/745022618356416572/c8fa739c82cdc5a730d9bdf411a552b0.png?size=1024",
         )
         embed.set_footer(text="Vous pouvez ajouter l'url du tournois, la commande ainsi que le graph du tournois dans un seul")
-        if not await self.ask_confirmation(ctx, embed):
+        if not await ask_confirmation(ctx, embed):
             await ctx.send("Commande annulée.")
             return
-        tournament_response = {
-            "tournament_event": {
-                "recurring_tournament_id": tournament["id"],
-                "bracket_url": bracket
-            }
+        tournament_data = {
+            "recurring_tournament_id": tournament["id"],
+            "bracket_url": bracket
         }
         if len(attachement) == 1:
-            tournament_response["tournament_event"]["graph_url"] = attachement[0].url
-        request_url = self.api_url("tournament_events")
-        async with self._session.post(request_url, json=tournament_response) as r:
-            if r.status == 201:
-                await self.show_confirmation(ctx, f"Une édition du tournois {tournament['name']} a été crée avec succès.")
-            elif r.status == 200:
-                await self.show_confirmation(ctx, f"Une édition du tournois {tournament['name']} a été modifié avec succès.")
-
-
-
-    # -------------------------------------------------------------------------
-    # COMMANDS
-    # -------------------------------------------------------------------------
-
-    @commands.command(usage="<nom>")
-    async def creerville(self, ctx, *, name):
-        """cette commande va vous permettre d'ajouter une ville dans la Smashthèque.
-        \n\nVous devez préciser son nom.
-        \n\n\nExemples : \n- creerville Paris\n- creerville Lyon\n"""
-
-        try:
-            await self.do_createlocation(ctx, name)
-        except:
-            rollbar.report_exc_info()
-            raise
-
-    @commands.command(usage="<nom>")
-    async def creerpays(self, ctx, *, name):
-        """cette commande va vous permettre d'ajouter un pays dans la Smashthèque.
-        \n\nVous devez préciser son nom.
-        \n\n\nExemples : \n- creerpays Belgique\n"""
-
-        try:
-            await self.do_createlocation(ctx, name, country=True)
-        except:
-            rollbar.report_exc_info()
-            raise
-
-    @commands.command(usage="<pseudo> <emojis ou noms de persos> [team] [localisation] [ID Discord]")
-    async def creerjoueur(self, ctx, *, arg):
-        """cette commande va vous permettre d'ajouter un joueur dans la Smashthèque.
-        \n\nVous devez ajouter au minimum le pseudo et les personnages joués (dans l'ordre).
-        \n\nVous pouvez aussi ajouter sa team, sa localisation et, s'il possède un compte Discord, son ID pour qu'il puisse modifier lui-même son compte.
-        \n\nVous pouvez récupérer l'ID avec les options de developpeur (activez-les dans l'onglet Apparence des paramètres de l'utilisateur, puis faites un clic droit sur l'utilisateur et sélectionnez \"Copier ID\".)
-        \n\n\nExemples : \n- creerjoueur Pixel <:Yoshi:737480513744273500> <:Bowser:737480497332224100>\n- creerjoueur Pixel Yoshi Bowser\n- creerjoueur red <:Joker:737480520052637756> LoS Paris 332894758076678144\n"""
-
-        try:
-            await self.do_addplayer(ctx, arg)
-        except:
-            rollbar.report_exc_info()
-            raise
-
-    @commands.command(usage="<pseudo>")
-    async def jesuis(self, ctx, *, pseudo):
-        """cette commande va vous permettre d'associer votre compte Discord à un joueur de la Smashthèque.
-        \n\nVous devez préciser un pseudo.
-        \n\n\nExemples : \n- jesuis Pixel\n- jesuis red\n"""
-
-        try:
-            await self.do_link(ctx, pseudo, ctx.author.id)
-        except:
-            rollbar.report_exc_info()
-            raise
-
-    @commands.command(usage="<pseudo> <ID Discord>")
-    @commands.admin_or_permissions(administrator=True)
-    async def associer(self, ctx, *, arg):
-        """cette commande va vous permettre d'associer un compte Discord à un joueur de la Smashthèque.
-        \n\nVous devez préciser son pseudo et son ID Discord.
-        \n\n\nExemples : \n- associer Pixel 608210202952466464\n- associer red 332894758076678144\n"""
-
-        try:
-            args = arg.split()
-            pseudo = ' '.join(args[:-1])
-            discord_id = args[-1]
-            await self.do_link(ctx, pseudo, discord_id)
-        except:
-            rollbar.report_exc_info()
-            raise
-
-    @commands.command()
-    @commands.admin_or_permissions(administrator=True)
-    async def jenesuispas(self, ctx):
-        """cette commande permet de dissocier votre compte Discord d'un joueur de la Smashthèque.
-        \n\n\nExemples : \n- jenesuispas\n"""
-
-        try:
-            await self.do_unlink(ctx, ctx.author)
-        except:
-            rollbar.report_exc_info()
-            raise
-
-    @commands.command(usage="<ID Discord>")
-    @commands.admin_or_permissions(administrator=True)
-    async def dissocier(self, ctx, *, target_member: discord.Member):
-        """cette commande permet de dissocier un compte Discord d'un joueur de la Smashthèque.
-        \n\nVous devez préciser son ID Discord.
-        \n\n\nExemples : \n- dissocier 608210202952466464\n- dissocier 332894758076678144\n"""
-
-        try:
-            await self.do_unlink(ctx, target_member)
-        except:
-            rollbar.report_exc_info()
-            raise
-
-    @commands.command(usage="<ID Discord>")
-    @commands.admin_or_permissions(administrator=True)
-    async def quiest(self, ctx, *, target_member: discord.Member):
-        try:
-            await self.do_showplayer(ctx, target_member)
-        except:
-            rollbar.report_exc_info()
-            raise
-
-    @commands.command()
-    async def quisuisje(self, ctx):
-        try:
-            await self.do_showplayer(ctx, ctx.author)
-        except:
-            rollbar.report_exc_info()
-            raise
-
-    @commands.command(usage="<nouveau pseudo>")
-    async def changerpseudo(self, ctx, *, name):
-        try:
-            await self.do_editname(ctx, ctx.author.id, name)
-        except:
-            rollbar.report_exc_info()
-            raise
-
-    @commands.command()
-    async def persos(self, ctx):
-        try:
-            await self.do_listavailablecharacters(ctx)
-        except:
-            rollbar.report_exc_info()
-            raise
-
-    @commands.command(usage="<emojis ou noms de persos>")
-    async def ajouterpersos(self, ctx, *, emojis):
-        try:
-            await self.do_addcharacters(ctx, ctx.author.id, emojis)
-        except:
-            rollbar.report_exc_info()
-            raise
-
-    @commands.command(usage="<emojis ou noms de persos>")
-    async def enleverpersos(self, ctx, *, emojis):
-        try:
-            await self.do_removecharacters(ctx, ctx.author.id, emojis)
-        except:
-            rollbar.report_exc_info()
-            raise
-
-    @commands.command(usage="<emojis ou noms de persos>")
-    async def remplacerpersos(self, ctx, *, emojis):
-        try:
-            await self.do_replacecharacters(ctx, ctx.author.id, emojis)
-        except:
-            rollbar.report_exc_info()
-            raise
-
-    @commands.command(usage="<team>")
-    async def quitter(self, ctx, *, team_short_name):
-        try:
-            await self.do_removeteam(ctx, ctx.author.id, team_short_name)
-        except:
-            rollbar.report_exc_info()
-            raise
-
-    @commands.command(usage="<team>")
-    async def integrer(self, ctx, *, team_short_name):
-        try:
-            await self.do_addteam(ctx, ctx.author.id, team_short_name)
-        except:
-            rollbar.report_exc_info()
-            raise
-
-    @commands.command(usage="<localisation>")
-    async def enleverlocalisation(self, ctx, *, location_name):
-        try:
-            await self.do_removelocation(ctx, ctx.author.id, location_name)
-        except:
-            rollbar.report_exc_info()
-            raise
-
-    @commands.command(usage="<localisation>")
-    async def ajouterlocalisation(self, ctx, *, location_name):
-        try:
-            await self.do_addlocation(ctx, ctx.author.id, location_name)
-        except:
-            rollbar.report_exc_info()
-            raise
-
-    @commands.command(usage="<pseudo>")
-    async def chercherjoueur(self, ctx, *, name):
-        try:
-            await self.do_findplayer(ctx, name)
-        except:
-            rollbar.report_exc_info()
-            raise
-
-    @commands.command()
-    async def pileouface(self, ctx):
-        try:
-            await self.do_chifoumi(ctx)
-        except:
-            rollbar.report_exc_info()
-            raise
-
-    @commands.command()
-    async def majlogo(self, ctx):
-        """Utilisez cette commande avec une image pour changer le logo de votre team. \n
-        Vous devez être administrateur de la team dont vous voulez changer le logo."""
-        try:
-            await self.maj_team_infos(ctx, "logo")
-        except:
-            rollbar.report_exc_info()
-            raise
-
-    @commands.command()
-    async def majroster(self, ctx):
-        """Utilisez cette commande avec une image pour changer le roster de votre team. \n
-        Vous devez être administrateur de la team dont vous voulez changer le roster."""
-        try:
-            await self.maj_team_infos(ctx, "roster")
-        except:
-            rollbar.report_exc_info()
-            raise
-
-    @commands.command()
-    async def tournoi(self, ctx, bracket: Optional[str]):
-        """Cette commande permet aux TOs d'ajouter une édition de leur tournois dans la smashtheque"""
-        try:
-            await self.do_addedition(ctx, bracket)
-        except:
-            rollbar.report_exc_info()
-            raise
+            tournament_data["graph_url"] = attachement[0].url
+        r = await self.api.createTournamentEvent(tournament_data)
+        if r.status == 201:
+            await show_confirmation(ctx, f"Une édition du tournois {tournament['name']} a été crée avec succès.")
+        elif r.status == 200:
+            await show_confirmation(ctx, f"Une édition du tournois {tournament['name']} a été modifié avec succès.")
